Many fixes to Jerusalem Bible.csv
Changed name to New Jerusalem Bible.csv, as I found out that it is in
fact the New Jerusalem Bible edition.
FIX: Formatting issues in numerous places where the CSV formatting was
not in place, leaving gaps between verses.
ADDED: Missing verses from the footnotes of the NJB (perhaps a bit too
liberally. Each addition is denoted with a * so that one would know it
is not in the plain text of the NJB (only in the footnotes).